pyproject.toml
[tool.poetry]
name = "proksee-batch"
version = "0.6.9"
description = "Proksee Batch"
authors = ["Lael D. Barlow"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/stothard-group/proksee-batch"
repository = "https://github.com/stothard-group/proksee-batch"
documentation = "https://proksee-batch.readthedocs.io"
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/stothard-group/proksee-batch/releases"

[tool.poetry.dependencies]
python = "~3.13"
click = ">=8.0.1"
biopython = ">=1.83"
requests = "^2.31.0"
selenium = "^4.16.0"
webdriver-manager = "^4.0.1"
types-selenium = "^3.141.9"
types-requests = "^2.31.0.20240106"
poetry-plugin-export = "^1.6.0"
toml = "^0.10.2"
types-toml = "^0.10.8.7"
gffutils = "^0.12"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
nox = ">=2022.11.21"
nox-poetry = ">=1.0.2"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyright = ">=1.1.350"
pyupgrade = ">=2.29.1"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}

[tool.poetry.scripts]
proksee-batch = "proksee_batch.__main__:main"

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/node_modules", "**/__pycache__", "**/.*"]
typeCheckingMode = "strict"
pythonVersion = "3.13"
pythonPlatform = "All"
reportMissingImports = true
reportMissingTypeStubs = false

# Set third-party related type errors to warning level
reportUnknownMemberType = "warning"
reportUnknownArgumentType = "warning"
reportUnknownVariableType = "warning"
reportUnknownParameterType = "warning"
reportUnknownLambdaType = "warning"
reportUnknownReturnType = "warning"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
